+++ src/main/java/com/example/hospitaldeliveryinterface/model/Employee.java
@@ -1,4 +1,4 @@
-package com.example.hospitaldeliveryinterface;
+package com.example.hospitaldeliveryinterface.model;
 
 public class Employee {
     private String firstName;
