--- src/main/resources/com/example/hospitaldeliveryinterface/Login.fxml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.hospitaldeliveryinterface.LoginController">
-   <children>
-      <ImageView fitHeight="682.0" fitWidth="1100.0">
-         <image>
-            <Image url="@../../../../../../../LogIn/src/main/resources/com/example/login/LogIn.jpg" />
-         </image>
-      </ImageView>
-      <AnchorPane layoutX="350.0" layoutY="151.0" prefHeight="386.0" prefWidth="410.0" style="-fx-background-color: #212121;">
-         <children>
-            <Label layoutX="155.0" layoutY="41.0" text="Login" textFill="WHITE">
-               <font>
-                  <Font name="Tahoma Bold" size="36.0" />
-               </font>
-            </Label>
-            <AnchorPane layoutX="170.0" layoutY="94.0" prefHeight="5.0" prefWidth="71.0" style="-fx-background-color: #76ff03;" />
-            <TextField fx:id="textFieldUsername" layoutX="110.0" layoutY="168.0" prefHeight="25.0" prefWidth="190.0" promptText="Username" />
-            <TextField fx:id="textFieldPassword" layoutX="111.0" layoutY="236.0" prefHeight="25.0" prefWidth="190.0" promptText="Password" />
-            <Button fx:id="LoginButton" layoutX="83.0" layoutY="291.0" mnemonicParsing="false" onAction="#handleLoginButton" prefHeight="46.0" prefWidth="245.0" style="-fx-background-color: #76ff03; -fx-background-radius: 90;" text="Login" textFill="WHITE">
-               <font>
-                  <Font name="Tahoma" size="14.0" />
-               </font>
-            </Button>
-            <ImageView fitHeight="38.0" fitWidth="33.0" layoutX="56.0" layoutY="162.0">
-               <image>
-                  <Image url="@../../../../../../../LogIn/src/main/resources/com/example/login/username.png" />
-               </image>
-            </ImageView>
-            <ImageView fitHeight="46.0" fitWidth="33.0" layoutX="50.0" layoutY="226.0">
-               <image>
-                  <Image url="@../../../../../../../LogIn/src/main/resources/com/example/login/password.png" />
-               </image>
-            </ImageView>
-         </children>
-      </AnchorPane>
-   </children>
-</AnchorPane>
