+++ src/main/resources/com/example/hospitaldeliveryinterface/OrderCard.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
-<GridPane fx:id="orderTemplate" alignment="CENTER_LEFT" hgap="20.0" maxHeight="-Infinity" minHeight="-Infinity" prefHeight="236.0" prefWidth="575.0" style="-fx-background-color: transparent; -fx-border-color: #22aae1; -fx-border-width: 2;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.hospitaldeliveryinterface.OrderCardUIController">
+<GridPane fx:id="orderTemplate" alignment="CENTER_LEFT" hgap="20.0" maxHeight="-Infinity" minHeight="-Infinity" prefHeight="236.0" prefWidth="575.0" style="-fx-background-color: transparent; -fx-border-color: #22aae1; -fx-border-width: 2;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.hospitaldeliveryinterface.controllers.OrderCardUIController">
   <columnConstraints>
     <ColumnConstraints halignment="LEFT" maxWidth="504.66666666666663" minWidth="10.0" prefWidth="119.33333333333331" />
     <ColumnConstraints halignment="LEFT" hgrow="SOMETIMES" maxWidth="484.6666323343913" minWidth="0.0" prefWidth="120.66672261555988" />
